More spacing around base characters with descenders next to medials
The lookups for adding extra guard space need to ignore narrow width non-spacing vowels that are above the base characters.
The list of such vowels needs to include all possible characters that would be found in a situation requiring extra guard space.
There may be more characters that needed to be added to this list. This list might be able to be merged with @classUpperVowel.

With this addition of another vowel, duplicates of this vowel in text will no longer be shown with a dotted circle,
but only in the DFLT shaper, not the much more commonly used mym2 shaper. If desired, such checking could be added to the mym2 lookups.

This issue was found in improving the selection of different forms of RA to handle the clashes (the medial is too close to the short RA) in the first three images at
https://github.com/silnrsi/font-padauk/issues/52#issuecomment-2259577005.
In the case of the forms of RA extra guard space might not be needed, the correct form of RA has the correct APs to position the medial correctly.

The next commit or so will improve the selection of RA forms.